test: reorganise coverage tests into topical modules
