+++ tests/tests_plotting/edgecases/test_plot_pytrendy_edgecases.py
@@ -140,3 +140,290 @@ class TestPlotPytrendyEdgeCases:
         self._prepare_and_plot(df, 'gradual', results.segments, suppress_show=False) # False, triggers plt.show()
         assert len(show_calls) == 1
 
+
+
+# =============================================================================
+# plot_pytrendy: boundary segments (first/last segment, non-neighbouring gaps)
+# =============================================================================
+
+class TestPlotBoundarySegments:
+    """Segment positioning at the edges of the index and gaps between segments."""
+
+    def test_first_segment_at_boundary_string(self):
+        """String index: first segment starts at index[0] (no prev)."""
+        df = pt.load_data('series_synthetic')
+        df['str_idx'] = [f'S{i}' for i in range(len(df))]
+        plot_df = df.set_index('str_idx')[['gradual']]
+        str_idx = list(plot_df.index)
+
+        segments = [
+            {'start': str_idx[0], 'end': str_idx[5], 'direction': 'Up',
+             'trend_class': 'gradual', 'change_rank': 1},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='string', suppress_show=True)
+        assert fig is not None
+        plt.close(fig)
+
+    def test_last_segment_at_boundary_string(self):
+        """String index: last segment ends at index[-1] (no next)."""
+        df = pt.load_data('series_synthetic')
+        df['str_idx'] = [f'S{i}' for i in range(len(df))]
+        plot_df = df.set_index('str_idx')[['gradual']]
+        str_idx = list(plot_df.index)
+
+        segments = [
+            {'start': str_idx[-10], 'end': str_idx[-1], 'direction': 'Down',
+             'trend_class': 'gradual', 'change_rank': 1},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='string', suppress_show=True)
+        assert fig is not None
+        plt.close(fig)
+
+    def test_non_neighbouring_segments_string(self):
+        """String index: segments with gaps (not adjacent)."""
+        df = pt.load_data('series_synthetic')
+        df['str_idx'] = [f'S{i}' for i in range(len(df))]
+        plot_df = df.set_index('str_idx')[['gradual']]
+        str_idx = list(plot_df.index)
+
+        segments = [
+            {'start': str_idx[1], 'end': str_idx[5], 'direction': 'Up',
+             'trend_class': 'gradual', 'change_rank': 1},
+            {'start': str_idx[20], 'end': str_idx[25], 'direction': 'Down',
+             'trend_class': 'gradual', 'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='string', suppress_show=True)
+        assert fig is not None
+        plt.close(fig)
+
+# =============================================================================
+# plot_pytrendy: plot customisation (plot_params branches)
+# =============================================================================
+
+class TestPlotCustomization:
+    """plot_params customisation branches (figsize/title/labels, legend, colours)."""
+
+    def _date_plot_df(self):
+        """Build a datetime-indexed DataFrame from synthetic data."""
+        df = pt.load_data('series_synthetic')
+        df['date'] = pd.to_datetime(df['date'])
+        return df.set_index('date')[['gradual']]
+
+    def test_plot_with_custom_params(self):
+        """Test plot_params path for date index type."""
+        plot_df = self._date_plot_df()
+        df = pt.load_data('series_synthetic')
+        results = pt.detect_trends(df, value_col='gradual', date_col='date',
+                                   plot=False, method_params={'abrupt_padding': 0})
+
+        plot_params = {
+            'figsize': (10, 3),
+            'title': 'Custom Title',
+            'xlabel': 'Custom X',
+            'ylabel': 'Custom Y',
+            'grid': {'visible': False},
+        }
+
+        fig = plot_pytrendy(plot_df, 'gradual', results.segments,
+                            index_type='date',
+                            suppress_show=True, plot_params=plot_params)
+        assert fig is not None
+        plt.close(fig)
+
+    def test_plot_with_custom_legend(self):
+        """Test legend customisation path."""
+        plot_df = self._date_plot_df()
+        df = pt.load_data('series_synthetic')
+        results = pt.detect_trends(df, value_col='gradual', date_col='date',
+                                   plot=False, method_params={'abrupt_padding': 0})
+
+        plot_params = {
+            'legend_loc': 'lower right',
+        }
+
+        fig = plot_pytrendy(plot_df, 'gradual', results.segments,
+                            index_type='date',
+                            suppress_show=True, plot_params=plot_params)
+        assert fig is not None
+        plt.close(fig)
+
+    def test_plot_with_custom_colors(self):
+        """Test custom colors path."""
+        plot_df = self._date_plot_df()
+        df = pt.load_data('series_synthetic')
+        results = pt.detect_trends(df, value_col='gradual', date_col='date',
+                                   plot=False, method_params={'abrupt_padding': 0})
+
+        plot_params = {
+            'colors': {'Up': 'lightgreen', 'Down': 'lightcoral'},
+        }
+
+        fig = plot_pytrendy(plot_df, 'gradual', results.segments,
+                            index_type='date',
+                            suppress_show=True, plot_params=plot_params)
+        assert fig is not None
+        plt.close(fig)
+
+
+# =============================================================================
+# plot_pytrendy: prev fill branch (lines 172-178, 182)
+# =============================================================================
+
+class TestPlotPrevFillDirect:
+    """Test the prev fill branch in plot_pytrendy when start displacement is invalid.
+
+    TODO: these examples use hand-crafted segment lists that are a bit contrived
+    to force the specific displacement conditions. Redo with more realistic
+    synthetic scenarios when a natural dataset produces these patterns.
+    """
+
+    @pytest.mark.plot
+    @pytest.mark.mpl_image_compare(baseline_dir='./',
+                                    filename='test_plot_string_prev_fill_direct.png',
+                                    style='default')
+    def test_string_prev_fill_direct(self):
+        """Lines 172-176, 182: string index, Flat→Up adjacent, invalid start displacement."""
+        # Custom data with a dip so Up start value < Flat end value
+        values = list(range(40))
+        values[19] = 25  # Flat end value (high)
+        values[20] = 15  # Up start value (low) — displacement invalid
+        df = pd.DataFrame({'date': [f'S{i}' for i in range(40)], 'gradual': values})
+        pt.detect_trends(df, date_col='date', value_col='gradual',
+                         plot=False, method_params={'abrupt_padding': 0})
+
+        # Craft segments: Flat S10-S19 (value 25 at end), adjacent Up S20-S35
+        # Up start value (15) < Flat end value (25) makes displacement invalid
+        str_idx = [f'S{i}' for i in range(40)]
+        plot_df = pd.DataFrame({'gradual': values}, index=str_idx)
+        segments = [
+            {'start': 'S10', 'end': 'S19', 'direction': 'Flat',
+             'change_rank': 1},
+            {'start': 'S20', 'end': 'S35', 'direction': 'Up',
+             'trend_class': 'gradual', 'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='string', suppress_show=True)
+        return fig
+
+    @pytest.mark.plot
+    @pytest.mark.mpl_image_compare(baseline_dir='./',
+                                    filename='test_plot_integer_prev_fill_direct.png',
+                                    style='default')
+    def test_integer_prev_fill_direct(self):
+        """Lines 177-178: integer index, Flat→Up adjacent, invalid start displacement."""
+        values = list(range(40))
+        values[19] = 25  # Flat end value (high)
+        values[20] = 15  # Up start value (low) — displacement invalid
+        df = pd.DataFrame({'gradual': values})
+        pt.detect_trends(df, value_col='gradual',
+                         plot=False, method_params={'abrupt_padding': 0})
+
+        plot_df = df[['gradual']]
+        segments = [
+            {'start': 10, 'end': 19, 'direction': 'Flat',
+             'change_rank': 1},
+            {'start': 20, 'end': 35, 'direction': 'Up',
+             'trend_class': 'gradual', 'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='integer', suppress_show=True)
+        return fig
+
+
+# =============================================================================
+# plot_pytrendy: next noise fill branch (lines 210-216)
+# =============================================================================
+
+class TestPlotNextNoiseFillDirect:
+    """Test the next noise fill branch in plot_pytrendy when end displacement is invalid.
+
+    TODO: these examples use hand-crafted segment lists that are a bit contrived
+    to force the specific displacement conditions. Redo with more realistic
+    synthetic scenarios when a natural dataset produces these patterns.
+    """
+
+    @pytest.mark.plot
+    @pytest.mark.mpl_image_compare(baseline_dir='./',
+                                    filename='test_plot_date_next_noise_fill_direct.png',
+                                    style='default')
+    def test_date_next_noise_fill_direct(self):
+        """Line 211: date index, Down→Noise adjacent, invalid end displacement."""
+        # Custom data where Down end value < next value (invalid for Down)
+        values = list(range(40))
+        values[24] = 10  # Down end value (low)
+        values[25] = 35  # Noise start value (high) — displacement invalid
+        df = pd.DataFrame({'date': pd.date_range('2025-01-01', periods=40, freq='D'),
+                           'gradual': values})
+        pt.detect_trends(df, date_col='date', value_col='gradual',
+                         plot=False, method_params={'abrupt_padding': 0})
+
+        plot_df = df.set_index('date')[['gradual']]
+        segments = [
+            {'start': pd.Timestamp('2025-01-02'), 'end': pd.Timestamp('2025-01-25'),
+             'direction': 'Down', 'trend_class': 'gradual', 'change_rank': 1},
+            {'start': pd.Timestamp('2025-01-26'), 'end': pd.Timestamp('2025-02-05'),
+             'direction': 'Noise', 'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='date', suppress_show=True)
+        return fig
+
+    @pytest.mark.plot
+    @pytest.mark.mpl_image_compare(baseline_dir='./',
+                                    filename='test_plot_string_next_noise_fill_direct.png',
+                                    style='default')
+    def test_string_next_noise_fill_direct(self):
+        """Lines 213-214: string index, Down→Noise adjacent, invalid end displacement."""
+        values = list(range(40))
+        values[24] = 10  # Down end value (low)
+        values[25] = 35  # Noise start value (high) — displacement invalid
+        df = pd.DataFrame({'date': [f'S{i}' for i in range(40)], 'gradual': values})
+        pt.detect_trends(df, date_col='date', value_col='gradual',
+                         plot=False, method_params={'abrupt_padding': 0})
+
+        str_idx = [f'S{i}' for i in range(40)]
+        plot_df = pd.DataFrame({'gradual': values}, index=str_idx)
+        segments = [
+            {'start': 'S1', 'end': 'S24', 'direction': 'Down',
+             'trend_class': 'gradual', 'change_rank': 1},
+            {'start': 'S25', 'end': 'S35', 'direction': 'Noise',
+             'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='string', suppress_show=True)
+        return fig
+
+    @pytest.mark.plot
+    @pytest.mark.mpl_image_compare(baseline_dir='./',
+                                    filename='test_plot_integer_next_noise_fill_direct.png',
+                                    style='default')
+    def test_integer_next_noise_fill_direct(self):
+        """Line 216: integer index, Down→Noise adjacent, invalid end displacement."""
+        values = list(range(40))
+        values[24] = 10  # Down end value (low)
+        values[25] = 35  # Noise start value (high) — displacement invalid
+        df = pd.DataFrame({'gradual': values})
+        pt.detect_trends(df, value_col='gradual',
+                         plot=False, method_params={'abrupt_padding': 0})
+
+        plot_df = df[['gradual']]
+        segments = [
+            {'start': 1, 'end': 24, 'direction': 'Down',
+             'trend_class': 'gradual', 'change_rank': 1},
+            {'start': 25, 'end': 35, 'direction': 'Noise',
+             'change_rank': 2},
+        ]
+
+        fig = plot_pytrendy(plot_df, 'gradual', segments,
+                            index_type='integer', suppress_show=True)
+        return fig
